feature: add unit test product repo

+++ utils/utils.go
@@ -31,7 +31,7 @@ func DumpIncomingContext(c context.Context) string {
 }
 
 // Offset to get offset from page and limit, min value for page = 1
-func Offset(page, limit int64) int64 {
+func Offset(page, limit int) int {
 	offset := (page - 1) * limit
 	if offset < 0 {
 		return 0
